Add NetworkTopology module.
This will be used to load files in from JSON, via a static function that
takes care of initialization and proper connections.

There is also a convenience function that will write out a given
NetworkTopology to JSON, so that you can create topologies easily in the
Python shell.

Code still needs to be tested since Flow isn't written yet.