Add initializer Python script and fix bugs.
See usage in initializer.py, or by just running the script with the -h
flag.

I've tested the initializer by creating a dummy subclass of Flow, and it
seems to work fine at outputting to JSON and reading from JSON (if you
run the script in the root directory). I will add test case 0 once we
have a better idea of what to do with Flow.