Set up Router functions and Events.
Router has been reorganized relative to the old design doc, and so has
RouterPacket. There are three hashmaps stored in Router now: a map from
neighbor to distance to that neighbor (based on link costs), a map from
RouterHost to the *min* cost needed to travel between the given Router
(which is a neighbor) and a known Host, and a map from Host to the *min*
cost to travel between this Router and that Host (even if that Host is
not directly connected). The last of these is actually used in
RouterPacket creation.