Fix bugs; Test Case 0 Reno "working".
Removed Packet ID buffer since it didn't seem to be necessary, so
retransmissions always go through even if the window size has been
decreased. In order to make sure that no new packets were being added
when the # of packets in transit was already too big, a modified check
has been added to FlowSendPacketsEvent. So the # of packets in transit
will not increase if it is > max window size.

Additionally, fixed bug in link.py where Link was instantaneously marked
as non-busy. Now, the Link will just check its status after a delay to
determine whether it is busy / whether it needs to send new events. This
leads to transmission times that do *not* exceed the link capacity.

Also changed MainEventLoop to just continue until num_flows_left = 0,
because there were some bugs I didn't notice before when that condition
wasn't set. We can undo this change later if we want (but we don't need
to).