Ignore out-of-date q delay data; TC1 Reno works.
Queuing delay data that is > 1 second old is considered out of date, and
hence ignored (see LinkBuffer).

Based on debugging, the flow does alternate between the "top branch" (L1
and L3) and the "bottom branch" (L2 and L4) of this topology (as
expected). Also, the Reno flow completes in around 26 seconds, whereas
the example posted on the course website completes in ~22 seconds.